+++ docs/development/CHANGELOG.md
@@ -7,6 +7,8 @@
 - [PR 24](https://github.com/OpenSenseAction/poligrain/pull/24) Allow to color
   CML paths when plotting on map using a cmap (by
   [@cchwala](https://github.com/cchwala))
+- [PR 29](https://github.com/OpenSenseAction/poligrain/pull/29) Add `xarray`
+  Accessor for plotting CML paths (by [@cchwala](https://github.com/cchwala))
 
 ### Bug fixes